Rename kind to fdio_vpp
