+++ nodes/fdio_vpp/clab_vpp.go.tpl
