+++ nodes/fdio_vpp/vpp_startup_config.go.tpl
